Add sign verify feature and unit tests
working tests

+++ universal-login-relayer/lib/http/routes/authorisation.ts
@@ -2,6 +2,13 @@ import {Router, Request} from 'express';
 import AuthorisationService, {AuthorisationRequest} from '../../integration/sql/services/authorisationService';
 import {asyncHandler, sanitize, responseOf, asString, asObject} from '@restless/restless';
 import {getDeviceInfo} from '../utils/getDeviceInfo';
+import {verifyCancelAuthroisationRequest, CancelAuthorisationRequest} from '@universal-login/commons';
+import { hashCancelAuthorisationRequest } from '@universal-login/commons/dist/lib/core/utils/authorisation';
+import {} from '@universal-login/contracts';
+import { ethers } from 'ethers';
+import WalletMasterWithRefund from '@universal-login/contracts/build/WalletMasterWithRefund.json';
+import { InvalidSignature, InvalidAddress } from '../../core/utils/errors';
+import { asSignature, MySignature } from '../utils/sanitizers';
 
 
 const request = (authorisationService : AuthorisationService) =>
@@ -17,13 +24,30 @@ const getPending = (authorisationService : AuthorisationService) =>
     return responseOf({ response: result });
   };
 
-const denyRequest = (authorisationService : AuthorisationService) =>
-  async (data: {walletContractAddress: string, body: {key: string}}) => {
+const denyRequest = (authorisationService : AuthorisationService, provider: any) =>
+  async (data: {walletContractAddress: string, body: {key: string, signedCancelAuthorisationRequest: MySignature}}) => {
+    const {walletContractAddress} = data;
+    const {key, signedCancelAuthorisationRequest: signature} = data.body;
+    const cancelAuthorisationRequest: CancelAuthorisationRequest = {walletContractAddress, key};
+    const payloadDigest = hashCancelAuthorisationRequest(cancelAuthorisationRequest);
+    const [isValid, computedAddress] = verifyCancelAuthroisationRequest(payloadDigest, signature, key);
+    if (!isValid) {
+      throw new InvalidSignature(`cancelAuthorisationRequest failed due to invalid signature`);
+    }
+
+    const contract = new ethers.Contract(walletContractAddress, WalletMasterWithRefund.interface, provider);
+    const flatSignature = ethers.utils.joinSignature(signature);
+    const isCorrectAddress = await contract.isValidSignature(payloadDigest, flatSignature);
+    if (!isCorrectAddress) {
+      throw new InvalidAddress(computedAddress);
+    }
+    console.log('isCorrentAddress', isCorrectAddress);
+
     const result = await authorisationService.removeRequest(data.walletContractAddress, data.body.key);
     return responseOf(result, 204);
   };
 
-export default (authorisationService : AuthorisationService) => {
+export default (authorisationService : AuthorisationService, provider: any) => {
   const router = Router();
 
   router.post('/', asyncHandler(
@@ -47,10 +71,11 @@ export default (authorisationService : AuthorisationService) => {
     sanitize({
       walletContractAddress: asString,
       body: asObject({
-        key: asString
+        key: asString,
+        signedCancelAuthorisationRequest: asSignature
       })
     }),
-    denyRequest(authorisationService)
+    denyRequest(authorisationService, provider)
   ));
 
   return router;
